requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
psutil
torchvision
git+https://github.com/pytorch/pytorch
torch
Pillow
